backtest: extract framework crate bridge
