+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
     ".",
     "crates/ploy-core",
     "crates/ploy-data",
+    "crates/ploy-backtest",
     "crates/ploy-polymarket",
     "crates/ploy-risk",
     "tools/sdk_auth_check",
@@ -22,6 +23,7 @@ ploy-core = { path = "crates/ploy-core" }
 ploy-data = { path = "crates/ploy-data", features = ["persistence"] }
 ploy-polymarket = { path = "crates/ploy-polymarket" }
 ploy-risk = { path = "crates/ploy-risk" }
+ploy-backtest = { path = "crates/ploy-backtest", features = ["persistence"] }
 
 # Async runtime
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "signal", "sync", "time", "net", "fs", "io-util", "process"] }
