+++ crates/ploy-backtest/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "ploy-backtest"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ploy-core = { path = "../ploy-core" }
+ploy-data = { path = "../ploy-data" }
+
+# Async runtime
+tokio = { version = "1", features = ["rt", "sync", "time"] }
+
+# Serialization
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+
+# Logging
+tracing = "0.1"
+
+# Numeric types
+rust_decimal = { version = "1", features = ["serde-with-str", "maths"] }
+rust_decimal_macros = "1"
+
+# Date/time
+chrono = { version = "0.4", features = ["serde"] }
+
+# Error handling
+anyhow = "1"
+
+# UUID (for recorder run IDs)
+uuid = { version = "1", features = ["v4", "serde"] }
+
+# CSV export
+csv = "1"
+
+# Database (optional, for DB-backed feeds and recorder)
+sqlx = { version = "0.8.6", default-features = false, features = ["runtime-tokio", "postgres", "chrono", "rust_decimal", "uuid", "macros", "json"], optional = true }
+
+[features]
+default = []
+persistence = ["dep:sqlx"]
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
