+++ crates/ploy-backtest/src/signal.rs
@@ -1,0 +1,70 @@
+//! Backtest signal types and strategy trait.
+//!
+//! Defines the `BacktestStrategy` trait that each backtest strategy must implement,
+//! along with signal and trade types used by the unified backtest engine.
+
+use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
+use serde::{Deserialize, Serialize};
+
+use crate::engine::MarketSnapshot;
+
+/// Signal generated by a backtest strategy.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Signal {
+    pub timestamp: DateTime<Utc>,
+    pub symbol: String,
+    pub direction: SignalDirection,
+    pub entry_price: Decimal,
+    pub shares: u64,
+    pub confidence: f64,
+    pub metadata: serde_json::Value,
+}
+
+/// Direction of a trading signal.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum SignalDirection {
+    BuyYes,
+    BuyNo,
+}
+
+/// Reason for exiting a trade.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum ExitReason {
+    /// Market resolved (binary settlement)
+    Settlement { outcome: bool },
+    /// Stop loss triggered
+    StopLoss,
+    /// Take profit triggered
+    TakeProfit,
+    /// Time-based exit (e.g., max holding period)
+    Timeout,
+    /// Strategy-specific exit
+    Custom(String),
+}
+
+/// An open trade being tracked by the engine.
+#[derive(Debug, Clone)]
+pub struct OpenTrade {
+    pub entry_time: DateTime<Utc>,
+    pub symbol: String,
+    pub direction: SignalDirection,
+    pub entry_price: Decimal,
+    pub shares: u64,
+    pub confidence: f64,
+}
+
+/// Trait that each backtest strategy must implement.
+pub trait BacktestStrategy: Send {
+    /// Strategy name for reporting.
+    fn name(&self) -> &str;
+
+    /// Process a market snapshot and optionally generate a signal.
+    fn on_snapshot(&mut self, snapshot: &MarketSnapshot) -> Option<Signal>;
+
+    /// Called when a trade is opened.
+    fn on_trade_open(&mut self, trade: &OpenTrade);
+
+    /// Check if an open trade should be closed.
+    fn check_exit(&mut self, trade: &OpenTrade, snapshot: &MarketSnapshot) -> Option<ExitReason>;
+}
