+++ crates/ploy-backtest/src/strategies/mod.rs
@@ -1,0 +1,4 @@
+//! Strategy implementations for backtesting.
+//!
+//! Concrete strategy implementations will be added here as they are
+//! migrated from the main application crate.
